==> notebooks/5.2 LLM Features - Tratamento.ipynb <==
# 5.2. Tratando as *Features* extraídas

* **Input**: Dados com as respostas de LLM não tratadas

* **Output**: Respostas de LLM tratas + variável com True/False da alternativa

## 5.2.1. Extração da tabela original

# Importando Dependências
import pandas as pd
import numpy as np
from scipy import stats
import statsmodels.api as sm

# Leitura dos dados
enem_data = pd.read_pickle("../data/final/enem_data_embeddings.pkl")
enem_data.head()

## 5.2.2. Extração das respostas do Llama

# leitura do csv

df_llama = pd.read_csv('../data/final/llama_extraction_csv.csv') 
df_llama.head()

# dropando coluna não utilizada

df_llama.drop('acerto_llamma', axis=1, inplace = True) 

# utilizando regex para padronizar respostas que a LLM deu

df_llama['resposta_llamma_limpo'] = df_llama['resposta_llamma'].str.findall(r'\b([A-E])\b').str[-1]

# gerando variável true/false de acertos do modelo Llama

df_llama["acerto_llamma"] = df_llama[["gabarito", "resposta_llamma_limpo"]].apply(
    lambda row: row["gabarito"] == row["resposta_llamma_limpo"], axis=1
)

# quantidade de acertos do Llama

df_llama["acerto_llamma"].sum()

# pegando apenas as colunas de interesse e as necessárias para o merge

df_llama = df_llama[['numero_questao','ANO','resposta_llamma_limpo','acerto_llamma']]

# padronizando nome

df_llama.rename(columns={'resposta_llamma_limpo': 'resposta_llama'}, inplace=True)


# verificando tabela final

df_llama.head()

## 5.2.3. Extração das respostas do Deepseek

# leitura do csv

df_deepseek = pd.read_csv('../data/final/deepseek_extraction_csv.csv') 
df_deepseek.head()

# dropando coluna não utilizada

df_deepseek.drop('acerto_deepseek', axis=1, inplace = True) 

# aqui visualizamos e tratamos manualmente algumas respostas mais complexas que o Deepseek forneceu

df_deepseek.loc[10, 'resposta_deepseek']

df_deepseek.loc[10, 'resposta_deepseek'] = 'B'

df_deepseek.loc[41, 'resposta_deepseek']

df_deepseek.loc[41, 'resposta_deepseek'] = 'C'

df_deepseek.loc[98, 'resposta_deepseek']

df_deepseek.loc[98, 'resposta_deepseek'] = 'C'

df_deepseek.loc[103, 'resposta_deepseek']

df_deepseek.loc[103, 'resposta_deepseek'] = 'E'

df_deepseek.loc[127, 'resposta_deepseek']

df_deepseek.loc[127, 'resposta_deepseek'] = 'C'

df_deepseek.loc[139, 'resposta_deepseek']

df_deepseek.loc[139, 'resposta_deepseek'] = 'E'

df_deepseek.loc[158, 'resposta_deepseek']

df_deepseek.loc[158, 'resposta_deepseek'] = 'C'



df_deepseek.loc[176, 'resposta_deepseek']

df_deepseek.loc[176, 'resposta_deepseek'] = 'B'

df_deepseek.loc[186, 'resposta_deepseek']

df_deepseek.loc[186, 'resposta_deepseek'] = 'B'

df_deepseek.loc[203, 'resposta_deepseek']

df_deepseek.loc[203, 'resposta_deepseek'] = 'E'

df_deepseek.loc[239, 'resposta_deepseek']

df_deepseek.loc[239, 'resposta_deepseek'] = 'E'

df_deepseek.loc[244, 'resposta_deepseek']

df_deepseek.loc[244, 'resposta_deepseek'] = 'C'

# agora, tratamos de forma mais geral as respostas menos complexas

df_deepseek['resposta_deepseek_limpo'] = df_deepseek['resposta_deepseek'].str.findall(r'\b([A-E])\b').str[-1]

# gerando variável de true/false de acertos do modelo Deepseek

df_deepseek["acerto_deepseek"] = df_deepseek[["gabarito", "resposta_deepseek_limpo"]].apply(
    lambda row: row["gabarito"] == row["resposta_deepseek_limpo"], axis=1
)

# quantidade de acertos do Deepseek

df_deepseek["acerto_deepseek"].sum()

# pegando apenas as colunas de interesse e as necessárias para o merge

df_deepseek = df_deepseek[['numero_questao','ANO','resposta_deepseek_limpo','acerto_deepseek']]

# padronizando nome

df_deepseek.rename(columns={'resposta_deepseek_limpo': 'resposta_deepseek'}, inplace=True)


# verificando tabela final

df_deepseek.head()

## 5.2.4. Extração das respostas do Deepseek

# inserindo as novas variáveis do Llama no dataset original

df_merged = pd.merge(
    enem_data,                 
    df_llama,                 
    on=['numero_questao', 'ANO'],  
    how='left'           
)

# inserindo as novas variáveis do Deepseek no dataset original

df_merged = pd.merge(
    df_merged,                 
    df_deepseek,                 
    on=['numero_questao', 'ANO'],  
    how='left'           
)

# resultado final

df_merged.head()

# salvando pickle
df_merged.to_pickle("../data/final/enem_data_final.pkl")
