==> notebooks/8. Random Forest.ipynb <==
# 8. Random Forest Regressor

Este notebook implementa um algoritmo de florestas de decisão adaptadas para regressão.. O objetivo é prever o parâmetro `nu_param_B` a partir dos textos dos enunciados. O fluxo inclui etapas de pré-processamento, tokenização, divisão dos dados, treinamento do modelo e avaliação dos resultados.

import pandas as pd
import numpy as np
import math
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, KFold, RandomizedSearchCV
from scipy.stats import randint
from scipy import stats
from sklearn.metrics import classification_report
from sklearn.metrics import mean_squared_error

## 8.1 Leitura da base e manipulação de variáveis

# Leitura do dataset com 50 dimensões
df = pd.read_pickle('../data/final/enem_data_final.pkl')
df.head()

# Manipulando variáveis
df_new = df[["enunciado_embbedings_word2vec_50",
               "acerto_deepseek",
                "acerto_llamma",
                "similaridade_enunciado_gabarito_50",
                "similaridade_enunciado_distratores_50",
                "similaridade_gabarito_distratores_50",
               'nu_param_B']]
df_new.rename(columns={'nu_param_B': 'y', "enunciado_embbedings_word2vec_50":'X'}, inplace=True)

# Removendo observações Nan
df_new = df_new[df_new['y'].notna()].reset_index(drop=True)

# Explodindo a coluna X em várias colunas
df_features = pd.DataFrame(df_new['X'].tolist(), index=df_new.index)
df_features.columns = [f'X{i+1}' for i in range(df_features.shape[1])]

# Criando dataset final
df_final = pd.concat([df_features, df_new.drop(['X', 'y'], axis=1)], axis=1)
df_final['acerto_llamma'] = df_final['acerto_llamma'].astype(float)
df_final['acerto_deepseek'] = df_final['acerto_deepseek'].astype(float)
df_final.head()

## 8.2 Modelagem

### 8.2.1 Separando dados

# Separando os dados
X = df_final
y = df_new['y']


# Aplicando Transformações
add_list = [(y.min() * (-1)) + 1] * len(y)
y = y + add_list

# Aplicando Boxcox
y, _ = stats.boxcox(y)
y = pd.Series(y)

# Dividir em treino e teste (para avaliação externa)
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42
)

### 8.2.2 Validação Cruzada

# Instancia o regressor
rf = RandomForestRegressor(random_state=42)

param_grid = {
    'n_estimators': [100, 200, 300],
    'max_depth': [None, 10, 20],
    'min_samples_split': [2, 5],
    'min_samples_leaf': [1, 2],
    'max_features': ['sqrt', 'log2']
}

# 'neg_mean_squared_error' retorna negativo do MSE
grid = GridSearchCV(
    estimator=rf,
    param_grid=param_grid,
    cv=5,
    scoring='neg_mean_squared_error',
    n_jobs=-1,
    verbose=1,
    refit=True
)

grid.fit(X_train, y_train)

best_rf = grid.best_estimator_
print("Melhores parâmetros:", grid.best_params_)

## 8.3 Avaliação

### 8.3.1 Conjunto de treino

cv = KFold(n_splits=5, shuffle=True, random_state=42)
mse_scores = -cross_val_score(
    best_rf, X_train, y_train,
    cv=cv,
    scoring='neg_mean_squared_error',
    n_jobs=-1
)
rmse_train_cv = np.sqrt(np.mean(mse_scores))
print(f"RMSE CV (treino): {rmse_train_cv:.4f}")

### 8.3.2 Conjunto de testes

y_pred = best_rf.predict(X_test)
mse_test = mean_squared_error(y_test, y_pred)
rmse_test = math.sqrt(mse_test)
print(f"RMSE (teste): {rmse_test:.4f}")
